Test format-seconds time formatter
